+++ priv/c/bendler_ask.js
@@ -1,0 +1,1 @@
+function bendler_ask() { throw new Error('bendler: the JS lane is not supported'); }